services/dashboard-service/requirements.txt
Flask==2.3.3
requests==2.31.0
paho-mqtt==1.6.1
